ai/twitter-scraper/requirements.txt
tweepy==4.16.0
requests==2.32.5
psycopg2-binary==2.9.10
anyio>=4.0.0
